services/result-service/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose[cryptography]
sqlalchemy
psycopg2-binary
pydantic